NMFList-class {NMF}	R Documentation

Class for Storing Heterogeneous NMF fits

Description

This class wraps a list of NMF fit objects, which may come from different runs of the function nmf, using different parameters, methods, etc.. These can be either from a single run (NMFfit) or multiple runs (NMFfitX).

Note that its definition/interface is very likely to change in the future.

Methods

algorithm

signature(object = "NMFList"): Returns the method names used to compute the NMF fits in the list. It returns NULL if the list is empty.

runtime

signature(object = "NMFList"): Returns the CPU time required to compute all NMF fits in the list. It returns NULL if the list is empty. If no timing data are available, the sequential time is returned.

seqtime

signature(object = "NMFList"): Returns the CPU time that would be required to sequentially compute all NMF fits stored in object.

This method calls the function runtime on each fit and sum up the results. It returns NULL on an empty object.

show

signature(object = "NMFList"): Show method for objects of class NMFList

[Package NMF version 0.21.0 Index]