NMFSeed-class {NMF}	R Documentation

Base class that defines the interface for NMF seeding methods.

Description

This class implements a simple wrapper strategy object that defines a unified interface to seeding methods, that are used to initialise NMF models before fitting them with any NMF algorithm.

Slots

name

character string giving the name of the seeding strategy

method

workhorse function that implements the seeding strategy. It must have signature (object="NMF", x="matrix", ...) and initialise the NMF model object with suitable values for fitting the target matrix x.

Methods

algorithm

signature(object = "NMFSeed"): Returns the workhorse function of the seeding method described by object.

algorithm<-

signature(object = "NMFSeed", value = "function"): Sets the workhorse function of the seeding method described by object.

NMFSeed

signature(key = "NMFSeed"): Creates an NMFSeed based on a template object (Constructor-Copy), in particular it uses the same name.

show

signature(object = "NMFSeed"): Show method for objects of class NMFSeed

[Package NMF version 0.21.0 Index]