NMFfitX {NMF}	R Documentation

Factory Method for Multiple NMF Run Objects

Description

Factory Method for Multiple NMF Run Objects

Usage

  NMFfitX(object, ...)

  ## S4 method for signature 'list'
NMFfitX(object, ..., .merge = FALSE)

Arguments

object
an object from which is created an NMFfitX object

...
extra arguments used to pass values for slots

.merge
a logical that indicates if the fits should be aggregated, only keeping the best fit, and return an NMFfitX1 object. If FALSE, an NMFfitXn object containing the data of all the fits is returned.

Methods

NMFfitX

signature(object = "list"): Create an NMFfitX object from a list of fits.

NMFfitX

signature(object = "NMFfit"): Creates an NMFfitX1 object from a single fit. This is used in nmf when only the best fit is kept in memory or on disk.

NMFfitX

signature(object = "NMFfitX"): Provides a way to aggregate NMFfitXn objects into an NMFfitX1 object.

[Package NMF version 0.21.0 Index]