getRNG1 {NMF}	R Documentation

Extracting RNG Data from NMF Objects

Description

The nmf function returns objects that contain embedded RNG data, that can be used to exactly reproduce any computation. These data can be extracted using dedicated methods for the S4 generics getRNG and getRNG1.

Usage

  getRNG1(object, ...)

  .getRNG(object, ...)

Arguments

object
an R object from which RNG settings can be extracted, e.g. an integer vector containing a suitable value for .Random.seed or embedded RNG data, e.g., in S3/S4 slot rng or rng$noise.

...
extra arguments to allow extension and passed to a suitable S4 method .getRNG or .setRNG.

Methods

.getRNG

signature(object = "NMFfitXn"): Returns the RNG settings used for the best fit.

This method throws an error if the object is empty.

getRNG1

signature(object = "NMFfitX"): Returns the RNG settings used for the first NMF run of multiple NMF runs.

getRNG1

signature(object = "NMFfitX1"): Returns the RNG settings used to compute the first of all NMF runs, amongst which object was selected as the best fit.

getRNG1

signature(object = "NMFfitXn"): Returns the RNG settings used for the first run.

This method throws an error if the object is empty.

Examples

# For multiple NMF runs, the RNG settings used for the first run is also stored
V <- rmatrix(20,10)
res <- nmf(V, 3, nrun=4)
# RNG used for the best fit
getRNG(res)
# RNG used for the first of all fits
getRNG1(res)
# they may differ if the best fit is not the first one
rng.equal(res, getRNG1(res))

[Package NMF version 0.21.0 Index]