Strategy-class {NMF}	R Documentation

Generic Strategy Class

Description

This class defines a common interface for generic algorithm strategies (e.g., NMFStrategy).

name and name<- gets and sets the name associated with an object. In the case of Strategy objects it is the the name of the algorithm.

Usage

  name(object, ...)

  ## S4 method for signature 'Strategy'
name(object, all = FALSE)

  name(object, ...)<-value

Arguments

object
an R object with a defined name method

...
extra arguments to allow extension

value
replacement value

all
a logical that indicates if all the names associated with a strategy should be returned (TRUE), or only the first (primary) one (FALSE).

Slots

name

character string giving the name of the algorithm

package

name of the package that defined the strategy.

defaults

default values for some of the algorithm's arguments.

Methods

name

signature(object = "Strategy"): Returns the name of an algorithm

name

signature(object = "Strategy"): Returns the name of an algorithm

name<-

signature(object = "Strategy", value = "character"): Sets the name(s) of an NMF algorithm

name<-

signature(object = "Strategy", value = "character"): Sets the name(s) of an NMF algorithm

[Package NMF version 0.21.0 Index]