advanced-NMF {NMF}	R Documentation

Advanced Usage of the Package NMF

Description

The functions documented here provide advanced functionalities useful when developing within the framework implemented in the NMF package.

which.best returns the index of the best fit in a list of NMF fit, according to some quantitative measure. The index of the fit with the lowest measure is returned.

Usage

  which.best(object, FUN = deviance, ...)

Arguments

object
an NMF model fitted by multiple runs.

FUN
the function that computes the quantitative measure.

...
extra arguments passed to FUN.

[Package NMF version 0.21.0 Index]