algorithm,NMFList-method {NMF}	R Documentation

Returns the method names used to compute the NMF fits in the list. It returns NULL if the list is empty.

Description

Returns the method names used to compute the NMF fits in the list. It returns NULL if the list is empty.

Usage

  ## S4 method for signature 'NMFList'
algorithm(object, string = FALSE,
    unique = TRUE)

Arguments

string
a logical that indicate whether the names should be collapsed into a comma-separated string.

unique
a logical that indicates whether the result should contain the set of method names, removing duplicated names. This argument is forced to TRUE when string=TRUE.

object
an object computed using some algorithm, or that describes an algorithm itself.

[Package NMF version 0.21.0 Index]