algorithm,NMFStrategyOctave-method {NMF}	R Documentation

Returns the name of the Octave/Matlab function that implements the NMF algorithm – as stored in slot algorithm.

Description

Returns the name of the Octave/Matlab function that implements the NMF algorithm – as stored in slot algorithm.

Usage

  ## S4 method for signature 'NMFStrategyOctave'
algorithm(object,
    load = FALSE)

Arguments

load
logical that indicates if the algorithm should be loaded as an R function.

object
an object computed using some algorithm, or that describes an algorithm itself.

[Package NMF version 0.20.6 Index]