.atrack {NMF}	R Documentation

Annotation Tracks

Description

.atrack is an S4 generic method that converts an object into an annotation track object. It provides a general and flexible annotation framework that is used by aheatmap to annotates heatmap rows and columns.

is.atrack tests if an object is an annotationTrack object.

adata get/sets the annotation parameters on an object

amargin get/sets the annotation margin, i.e. along which dimension of the data the annotations are to be considered.

anames returns the reference margin names for annotation tracks, from their embedded annotation data object.

alength returns the reference length for annotation tracks, from their embedded annotation data object

atrack creates/concatenates annotationTrack objects

annotationTrack is constructor function for annotationTrack object

Usage

  .atrack(object, ...)

  is.atrack(x)

  adata(x, value, ...)

  amargin(x, value)

  anames(x, default.margin)

  alength(x, default.margin)

  ## S4 method for signature 'ANY'
.atrack(object, data = NULL, ...)

  atrack(..., order = NULL, enforceNames = FALSE,
    .SPECIAL = NA, .DATA = NULL, .CACHE = NULL)

  annotationTrack(x = list())

Arguments

object
an object from which is extracted annotation tracks

...
extra arguments to allow extensions and passed to the next method call. For atrack, arguments in ... are concatenated into a single annotationTrack object.

x
an R object

value
replacement value for the complete annotation data list

default.margin
margin to use if no margin data is stored in the x.

data
object used to extend the annotation track within a given data context. It is typically a matrix-like object, against which annotation specifications are matched using match_atrack.

order
an integer vector that indicates the order of the annotation tracks in the result list

enforceNames
logical that indicates if missing track names should be generated as X<i>

.SPECIAL
an optional list of functions (with no arguments) that are called to generate special annotation tracks defined by codes of the form ':NAME'. e.g., the function link{consensusmap} defines special tracks ':basis' and ':consensus'.

If .SPECIAL=FALSE, then any special tracks is discarded and a warning is thrown.

.DATA
data used to match and extend annotation specifications. It is passed to argument data of the .atrack methods, which in turn use pass it to match_atrack.

.CACHE
an annotationTrack object with which the generated annotation track should be consistent. This argument is more for internal/advanced usage and should not be used by the end-user.

Details

Methods for .atrack exist for common type of objects, which should provide enough options for new methods to define how annotation track are extracted from more complex objects, by coercing/filtering them into a supported type.

Value

atrack returns a list, decorated with class 'annotationTrack', where each element contains the description of an annotation track.

Methods

.atrack

signature(object = "ANY"): The default method converts character or integer vectors into factors. Numeric vectors, factors, a single NA or annotationTrack objects are returned unchanged (except from reordering by argument order). Data frames are not changed either, but class 'annotationTrack' is appended to their original class set.

[Package NMF version 0.21.0 Index]