bioc-NMF {NMF}	R Documentation

Specific NMF Layer for Bioconductor

Description

The package NMF provides an optional layer for working with common objects and functions defined in the Bioconductor platform.

Details

It provides:

computation functions that support ExpressionSet objects as inputs.

aliases and methods for generic functions defined and widely used by Bioconductor base packages.

specialised visualisation methods that adapt the titles and legend using bioinformatics terminology.

functions to link the results with annotations, etc...

[Package NMF version 0.21.0 Index]