c,NMF-method {NMF}	R Documentation

Concatenating NMF Models

Description

Binds compatible matrices and NMF models together.

Usage

  ## S4 method for signature 'NMF'
c(x, ..., margin = 3, recursive = FALSE)

Arguments

x
an NMF model

...
other objects to concatenate. Currently only two objects at a time can be concatenated (i.e. x and ..1).

margin
integer that indicates the margin along which to concatenate (only used when ..1 is a matrix):

1L

2L

3L

4L

If missing the margin is heuristically determined by looking at common dimensions between the objects.

recursive
logical. If recursive = TRUE, the function recursively descends through lists (and pairlists) combining all their elements into a vector.

[Package NMF version 0.21.0 Index]