canFit {NMF}	R Documentation

Testing Compatibility of Algorithm and Models

Description

canFit is an S4 generic that tests if an algorithm can fit a particular model.

Usage

  canFit(x, y, ...)

  ## S4 method for signature 'NMFStrategy,character'
canFit(x, y,
    exact = FALSE)

Arguments

x
an object that describes an algorithm

y
an object that describes a model

...
extra arguments to allow extension

exact
for logical that indicates if an algorithm is considered able to fit only the models that it explicitly declares (TRUE), or if it should be considered able to also fit models that extend models that it explicitly fits.

Methods

canFit

signature(x = "NMFStrategy", y = "character"): Tells if an NMF algorithm can fit a given class of NMF models

canFit

signature(x = "NMFStrategy", y = "NMF"): Tells if an NMF algorithm can fit the same class of models as y

canFit

signature(x = "character", y = "ANY"): Tells if a registered NMF algorithm can fit a given NMF model

See Also

Other regalgo: nmfAlgorithm

[Package NMF version 0.21.0 Index]