cluster_mat {NMF}	R Documentation

Cluster Matrix Rows in Annotated Heatmaps

Description

Cluster Matrix Rows in Annotated Heatmaps

Usage

  cluster_mat(mat, param, distfun, hclustfun, reorderfun,
    na.rm = TRUE, subset = NULL, verbose = FALSE)

Arguments

mat
original input matrix that has already been appropriately subset in the caller function (aheatmap)

param
clustering specifications

distfun
Default distance method/function

hclustfun
Default clustering (linkage) method/function

reorderfun
Default reordering function

na.rm
Logical that specifies if NA values should be removed

subset
index (integer) vector specifying the subset indexes used to subset mat. This is required to be able to return the original indexes.

[Package NMF version 0.21.0 Index]