consensus,NMFfitX1-method {NMF}	R Documentation

Returns the consensus matrix computed while performing all NMF runs, amongst which object was selected as the best fit.

Description

The result is the matrix stored in slot ‘consensus’. This method returns NULL if the consensus matrix is empty.

Usage

  ## S4 method for signature 'NMFfitX1'
consensus(object, no.attrib = FALSE)

Arguments

object
an object with a suitable predict method.

no.attrib
a logical that indicates if attributes containing information about the NMF model should be attached to the result (TRUE) or not (FALSE).

[Package NMF version 0.21.0 Index]