consensus,NMFfitXn-method {NMF}	R Documentation

Computes the consensus matrix of the set of fits stored in object, as the mean connectivity matrix across runs.

Description

This method returns NULL on an empty object. The result is a matrix with several attributes attached, that are used by plotting functions such as consensusmap to annotate the plots.

Usage

  ## S4 method for signature 'NMFfitXn'
consensus(object, ...,
    no.attrib = FALSE)

Arguments

object
an object with a suitable predict method.

...
extra arguments to allow extension. They are passed to predict, except for the vector and factor methods.

no.attrib
a logical that indicates if attributes containing information about the NMF model should be attached to the result (TRUE) or not (FALSE).

[Package NMF version 0.21.0 Index]