consensushc {NMF}	R Documentation

Hierarchical Clustering of a Consensus Matrix

Description

The function consensushc computes the hierarchical clustering of a consensus matrix, using the matrix itself as a similarity matrix and average linkage. It is

Usage

  consensushc(object, ...)

  ## S4 method for signature 'matrix'
consensushc(object,
    method = "average", dendrogram = TRUE)

  ## S4 method for signature 'NMFfitX'
consensushc(object,
    what = c("consensus", "fit"), ...)

Arguments

object
a matrix or an NMFfitX object, as returned by multiple NMF runs.

...
extra arguments passed to next method calls

method
linkage method passed to hclust.

dendrogram
a logical that specifies if the result of the hierarchical clustering (en hclust object) should be converted into a dendrogram. Default value is TRUE.

what
character string that indicates which matrix to use in the computation.

Value

an object of class dendrogram or hclust depending on the value of argument dendrogram.

Methods

consensushc

signature(object = "matrix"): Workhorse method for matrices.

consensushc

signature(object = "NMF"): Compute the hierarchical clustering on the connectivity matrix of object.

consensushc

signature(object = "NMFfitX"): Compute the hierarchical clustering on the consensus matrix of object, or on the connectivity matrix of the best fit in object.

[Package NMF version 0.21.0 Index]