cophcor {NMF}	R Documentation

Cophenetic Correlation Coefficient

Description

The function cophcor computes the cophenetic correlation coefficient from consensus matrix object, e.g. as obtained from multiple NMF runs.

Usage

  cophcor(object, ...)

  ## S4 method for signature 'matrix'
cophcor(object, linkage = "average")

Arguments

object
an object from which is extracted a consensus matrix.

...
extra arguments to allow extension and passed to subsequent calls.

linkage
linkage method used in the hierarchical clustering. It is passed to hclust.

Details

The cophenetic correlation coeffificient is based on the consensus matrix (i.e. the average of connectivity matrices) and was proposed by Brunet et al. (2004) to measure the stability of the clusters obtained from NMF.

It is defined as the Pearson correlation between the samples' distances induced by the consensus matrix (seen as a similarity matrix) and their cophenetic distances from a hierachical clustering based on these very distances (by default an average linkage is used). See Brunet et al. (2004).

Methods

cophcor

signature(object = "matrix"): Workhorse method for matrices.

cophcor

signature(object = "NMFfitX"): Computes the cophenetic correlation coefficient on the consensus matrix of object. All arguments in ... are passed to the method cophcor,matrix.

References

Brunet J, Tamayo P, Golub TR and Mesirov JP (2004). "Metagenes and molecular pattern discovery using matrix factorization." _Proceedings of the National Academy of Sciences of the United States of America_, *101*(12), pp. 4164-9. ISSN 0027-8424, <URL: http://dx.doi.org/10.1073/pnas.0308531101>, <URL: http://www.ncbi.nlm.nih.gov/pubmed/15016911>.

See Also

cophenetic

[Package NMF version 0.21.0 Index]