tryViewport {NMF}	R Documentation

Internal Grid Extension

Description

These functions enable mixing base and grid graphics in aheatmap, by avoiding calls to the grid internal function 'L_gridDirty'. They are not exported (i.e. not tampering core functions) and are only meant for internal use within the NMF package.

tryViewport tries to go down to a viewport in the current tree, given its name.

current.vpPath_patched aims at substituting current.vpPath, so that the graphic engine is not reset. This is essentially to prevent outputting a blank page at the beginning of PDF graphic engines.

.use.grid.patch tells if the user enabled patching grid.

Usage

  tryViewport(name, verbose = FALSE)

  current.vpPath_patched()

  .use.grid.patch()

Arguments

name
viewport name

verbose
toggle verbosity

Details

tryViewport uses grid.ls and not seekViewport as the latter would reset the graphic device and break the mix grid/base graphic capability.

[Package NMF version 0.21.0 Index]