pmax.inplace {NMF}	R Documentation

Updating Objects In Place

Description

These functions modify objects (mainly matrix objects) in place, i.e. they act directly on the C pointer. Due to their side-effect, they are not meant to be called by the end-user.

neq.constraints.inplace apply unequality constraints in place.

Usage

  pmax.inplace(x, lim, skip = NULL)

  neq.constraints.inplace(x, constraints, ratio = NULL,
    value = NULL, copy = FALSE)

Arguments

x
an object to update in place.

lim
lower threshold value

skip
indexes to skip

constraints
constraint specification.

ratio
fixed ratio on which the constraint applies.

value
fixed value to enforce.

copy
a logical that indicates if x should be updated in place or not.

Details

pmax.inplace is a version of pmax that updates its first argument.

[Package NMF version 0.21.0 Index]