nmfCheck {NMF}	R Documentation

Checking NMF Algorithm

Description

nmfCheck enables to quickly check that a given NMF algorithm runs properly, by applying it to some small random data.

Usage

  nmfCheck(method = NULL, rank = max(ncol(x)/5, 3),
    x = NULL, seed = 1234, ...)

Arguments

method
name of the NMF algorithm to be tested.

rank
rank of the factorization

x
target data. If NULL, a random 20 x 10 matrix is generated

seed
specifies a seed or seeding method for the computation.

...
other arguments passed to the call to nmf.

Value

the result of the NMF fit invisibly.

Examples

# test default algorithm
nmfCheck()

# test 'lee' algorithm
nmfCheck('lee')

[Package NMF version 0.21.0 Index]