nmfObject {NMF}	R Documentation

Updating NMF Objects

Description

This function serves to update an objects created with previous versions of the NMF package, which would otherwise be incompatible with the current version, due to changes in their S4 class definition.

Usage

  nmfObject(object, verbose = FALSE)

Arguments

object
an R object created by the NMF package, e.g., an object of class NMF or NMFfit.

verbose
logical to toggle verbose messages.

Details

This function makes use of heuristics to automatically update object slots, which have been borrowed from the BiocGenerics package, the function updateObjectFromSlots in particular.

[Package NMF version 0.21.0 Index]