nmfReport {NMF}	R Documentation

Run NMF Methods and Generate a Report

Description

Generates an HTML report from running a set of method on a given target matrix, for a set of factorization ranks.

Usage

  nmfReport(x, rank, method, colClass = NULL, ...,
    output = NULL, template = NULL)

Arguments

x
target matrix

rank
factorization rank

method
list of methods to apply

colClass
reference class to assess accuracy

...
extra paramters passed to nmf

output
output HTML file

template
template Rmd file

Details

The report is based on an .Rmd document 'report.Rmd' stored in the package installation sub-directory scripts/, and is compiled using knitr.

At the beginning of the document, a file named 'functions.R' is looked for in the current directory, and sourced if present. This enables the definition of custom NMF methods (see setNMFMethod) or setting global options.

Value

a list with the following elements:

fits
the fit(s) for each method and each value of the rank.

accuracy
a data.frame that contains the summary assessment measures, for each fit.

Examples

## Not run: 

x <- rmatrix(20, 10)
gr <- gl(2, 5)
nmfReport(x, 2:4, method = list('br', 'lee'), colClass = gr, nrun = 5)

## End(Not run)

[Package NMF version 0.21.0 Index]