objective,NMFfit-method {NMF}	R Documentation

Returns the objective function associated with the algorithm that computed the fitted NMF model object, or the objective value with respect to a given target matrix y if it is supplied.

Description

Returns the objective function associated with the algorithm that computed the fitted NMF model object, or the objective value with respect to a given target matrix y if it is supplied.

Usage

  ## S4 method for signature 'NMFfit'
objective(object, y)

Arguments

y
optional target matrix used to compute the objective value.

object
an object computed using some algorithm, or that describes an algorithm itself.

[Package NMF version 0.21.0 Index]