options-NMF {NMF}	R Documentation

NMF Package Specific Options

Description

NMF Package Specific Options

nmf.options sets/get single or multiple options, that are specific to the NMF package. It behaves in the same way as options.

nmf.getOption returns the value of a single option, that is specific to the NMF package. It behaves in the same way as getOption.

nmf.resetOptions reset all NMF specific options to their default values.

nmf.printOptions prints all NMF specific options along with their default values, in a relatively compact way.

Usage

  nmf.options(...)

  nmf.getOption(x, default = NULL)

  nmf.resetOptions(..., ALL = FALSE)

  nmf.printOptions()

Arguments

...
option specifications. For nmf.options this can be named arguments or a single unnamed argument that is a named list (see options.

For nmf.resetOptions, this must be the names of the options to reset. Note that pkgmaker version >= 0.9.1 is required for this to work correctly, when options other than the default ones have been set after the package is loaded.

ALL
logical that indicates if options that are not part of the default set of options should be removed. Note that in pkgmaker <= 0.9 this argument is only taken into account when no other argument is present. This is fixed in version 0.9.1.

x
a character string holding an option name.

default
if the specified option is not set in the options list, this value is returned. This facilitates retrieving an option and checking whether it is set and setting it separately if not.

Available options

cores

Default number of cores to use to perform parallel NMF computations. Note that this option is effectively used only if the global option 'cores' is not set. Moreover, the number of cores can also be set at runtime, in the call to nmf, via arguments .pbackend or .options (see nmf for more details).

default.algorithm

Default NMF algorithm used by the nmf function when argument method is missing. The value should the key of one of the registered NMF algorithms or a valid specification of an NMF algorithm. See ?nmfAlgorithm.

default.seed

Default seeding method used by the nmf function when argument seed is missing. The value should the key of one of the registered seeding methods or a vallid specification of a seeding method. See ?nmfSeed.

track

Toggle default residual tracking. When TRUE, the nmf function compute and store the residual track in the result – if not otherwise specified in argument .options. Note that tracking may significantly slow down the computations.

track.interval

Number of iterations between two points in the residual track. This option is relevant only when residual tracking is enabled. See ?nmf.

error.track

this is a symbolic link to option track for backward compatibility.

pbackend

Default loop/parallel foreach backend used by the nmf function when argument .pbackend is missing. Currently the following values are supported: 'par' for multicore, 'seq' for sequential, NA for standard sapply (i.e. do not use a foreach loop), NULL for using the currently registered foreach backend.

parallel.backend

this is a symbolic link to option pbackend for backward compatibility.

gc

Interval/frequency (in number of runs) at which garbage collection is performed.

verbose

Default level of verbosity.

debug

Toogles debug mode. In this mode the console output may be very – very – messy, and is aimed at debugging only.

maxIter

Default maximum number of iteration to use (default NULL). This option is for internal/technical usage only, to globally speed up examples or tests of NMF algorithms. To be used with care at one's own risk... It is documented here so that advanced users are aware of its existence, and can avoid possible conflict with their own custom options.

Examples

# show all NMF specific options
nmf.printOptions()

# get some options
nmf.getOption('verbose')
nmf.getOption('pbackend')
# set new values
nmf.options(verbose=TRUE)
nmf.options(pbackend='mc', default.algorithm='lee')
nmf.printOptions()

# reset to default
nmf.resetOptions()
nmf.printOptions()

[Package NMF version 0.21.0 Index]