plot,NMFfit,missing-method {NMF}	R Documentation

Plots the residual track computed at regular interval during the fit of the NMF model x.

Description

Plots the residual track computed at regular interval during the fit of the NMF model x.

Usage

  ## S4 method for signature 'NMFfit,missing'
plot(x, y, skip = -1, ...)

Arguments

skip
an integer that indicates the number of points to skip/remove from the beginning of the curve. If skip=1L (default) only the initial residual – that is computed before any iteration, is skipped, if present in the track (it associated with iteration 0).

x
the coordinates of points in the plot. Alternatively, a single plotting structure, function or any R object with a plot method can be provided.

y
the y coordinates of points in the plot, optional if x is an appropriate structure.

...
Arguments to be passed to methods, such as graphical parameters (see par). Many methods will accept the following arguments:

type

what type of plot should be drawn. Possible types are

"p" for points,

"l" for lines,

"b" for both,

"c" for the lines part alone of "b",

"o" for both ‘overplotted’,

"h" for ‘histogram’ like (or ‘high-density’) vertical lines,

"s" for stair steps,

"S" for other steps, see ‘Details’ below,

"n" for no plotting.

All other types give a warning or an error; using, e.g., type = "punkte" being equivalent to type = "p" for S compatibility. Note that some methods, e.g. plot.factor, do not accept this.

main

an overall title for the plot: see title.

sub

a sub title for the plot: see title.

xlab

a title for the x axis: see title.

ylab

a title for the y axis: see title.

asp

the y/x aspect ratio, see plot.window.

[Package NMF version 0.21.0 Index]