randomize {NMF}	R Documentation

Randomizing Data

Description

randomize permutates independently the entries in each column of a matrix-like object, to produce random data that can be used in permutation tests or bootstrap analysis.

Usage

  randomize(x, ...)

Arguments

x
data to be permutated. It must be an object suitable to be passed to the function apply.

...
extra arguments passed to the function sample.

Details

In the context of NMF, it may be used to generate random data, whose factorization serves as a reference for selecting a factorization rank, that does not overfit the data.

Value

a matrix

Examples

x <- matrix(1:32, 4, 8)
randomize(x)
randomize(x)

[Package NMF version 0.21.0 Index]