methods-NMF {NMF}	R Documentation

Registry for NMF Algorithms

Description

Registry for NMF Algorithms

selectNMFMethod tries to select an appropriate NMF algorithm that is able to fit a given the NMF model.

getNMFMethod retrieves NMF algorithm objects from the registry.

existsNMFMethod tells if an NMF algorithm is registered under the

removeNMFMethod removes an NMF algorithm from the registry.

Usage

  selectNMFMethod(name, model, load = FALSE, exact = FALSE,
    all = FALSE, quiet = FALSE)

  getNMFMethod(...)

  existsNMFMethod(name, exact = TRUE)

  removeNMFMethod(name, ...)

Arguments

name
name of a registered NMF algorithm

model
class name of an NMF model, i.e. a class that inherits from class NMF.

load
a logical that indicates if the selected algorithms should be loaded into NMFStrategy objects

all
a logical that indicates if all algorithms that can fit model should be returned or only the default or first found.

quiet
a logical that indicates if the operation should be performed quietly, without throwing errors or warnings.

...
extra arguments passed to pkgreg_fetch or pkgreg_remove.

exact
a logical that indicates if the access key should be matched exactly (TRUE) or partially (FALSE).

Value

selectNMFMethod returns a character vector or NMFStrategy objects, or NULL if no suitable algorithm was found.

[Package NMF version 0.21.0 Index]