runtime,NMFList-method {NMF}	R Documentation

Returns the CPU time required to compute all NMF fits in the list. It returns NULL if the list is empty. If no timing data are available, the sequential time is returned.

Description

Returns the CPU time required to compute all NMF fits in the list. It returns NULL if the list is empty. If no timing data are available, the sequential time is returned.

Usage

  ## S4 method for signature 'NMFList'
runtime(object, all = FALSE)

Arguments

all
logical that indicates if the CPU time of each fit should be returned (TRUE) or only the total CPU time used to compute all the fits in object.

object
an object computed using some algorithm, or that describes an algorithm itself.

[Package NMF version 0.21.0 Index]