runtime.all,NMFfitXn-method {NMF}	R Documentation

Returns the CPU time used to perform all the NMF fits stored in object.

Description

If no time data is available from in slot ‘runtime.all’ and argument null=TRUE, then the sequential time as computed by seqtime is returned, and a warning is thrown unless warning=FALSE.

Usage

  ## S4 method for signature 'NMFfitXn'
runtime.all(object, null = FALSE,
    warning = TRUE)

Arguments

null
a logical that indicates if the sequential time should be returned if no time data is available in slot ‘runtime.all’.

warning
a logical that indicates if a warning should be thrown if the sequential time is returned instead of the real CPU time.

object
an object computed using some algorithm, or that describes an algorithm itself.

[Package NMF version 0.21.0 Index]