staticVar {NMF}	R Documentation

Get/Set a Static Variable in NMF Algorithms

Description

This function is used in iterative NMF algorithms to manage variables stored in a local workspace, that are accessible to all functions that define the iterative schema described in NMFStrategyIterative.

It is specially useful for computing stopping criteria, which often require model data from different iterations.

Usage

  staticVar(name, value, init = FALSE)

Arguments

name
Name of the static variable (as a single character string)

value
New value of the static variable

init
a logical used when a value is provided, that specifies if the variable should be set to the new value only if it does not exist yet (init=TRUE).

Value

The value of the static variable

[Package NMF version 0.21.0 Index]