t.NMF {NMF}	R Documentation

Transformation NMF Model Objects

Description

t transpose an NMF model, by transposing and swapping its basis and coefficient matrices: t([W,H]) = [t(H), t(W)].

Usage

  ## S3 method for class 'NMF'
 t(x)

Arguments

x
NMF model object.

Details

The function t is a generic defined in the base package. The method t.NMF defines the trasnformation for the general NMF interface. This method may need to be overloaded for NMF models, whose structure requires specific handling.

See Also

Other transforms: nneg, posneg, rposneg

Examples

x <- rnmf(3, 100, 20)
x
# transpose
y <- t(x)
y

# factors are swapped-transposed
stopifnot( identical(basis(y), t(coef(x))) )
stopifnot( identical(coef(y), t(basis(x))) )

[Package NMF version 0.21.0 Index]