ibterms {NMF}	R Documentation

Fixed Terms in NMF Models

Description

Formula-based NMF models may contain fixed basis and/or coefficient terms. The functions documented here provide access to these data, which are read-only and defined when the model object is instantiated (e.g., see nmfModel,formula-method).

ibterms, icterms and iterms respectively return the indexes of the fixed basis terms, the fixed coefficient terms and all fixed terms, within the basis and/or coefficient matrix of an NMF model.

nterms, nbterms, and ncterms return, respectively, the number of all fixed terms, fixed basis terms and fixed coefficient terms in an NMF model. In particular: i.e. nterms(object) = nbterms(object) + ncterms(object).

bterms and cterms return, respectively, the primary data for fixed basis and coefficient terms in an NMF model – as stored in slots bterms and cterms . These are factors or numeric vectors which define fixed basis components, e.g., used for defining separate offsets for different a priori groups of samples, or to incorporate/correct for some known covariate.

ibasis and icoef return, respectively, the indexes of all latent basis vectors and estimated coefficients within the basis or coefficient matrix of an NMF model.

Usage

  ibterms(object, ...)

  icterms(object, ...)

  iterms(object, ...)

  nterms(object)

  nbterms(object)

  ncterms(object)

  bterms(object)

  cterms(object)

  ibasis(object, ...)

  icoef(object, ...)

Arguments

object
NMF object

...
extra parameters to allow extension (currently not used)

Methods

ibterms

signature(object = "NMF"): Default pure virtual method that ensure a method is defined for concrete NMF model classes.

ibterms

signature(object = "NMFstd"): Method for standard NMF models, which returns the integer vector that is stored in slot ibterms when a formula-based NMF model is instantiated.

ibterms

signature(object = "NMFfit"): Method for single NMF fit objects, which returns the indexes of fixed basis terms from the fitted model.

ibterms

signature(object = "NMFfitX"): Method for multiple NMF fit objects, which returns the indexes of fixed basis terms from the best fitted model.

icterms

signature(object = "NMF"): Default pure virtual method that ensure a method is defined for concrete NMF model classes.

icterms

signature(object = "NMFstd"): Method for standard NMF models, which returns the integer vector that is stored in slot icterms when a formula-based NMF model is instantiated.

icterms

signature(object = "NMFfit"): Method for single NMF fit objects, which returns the indexes of fixed coefficient terms from the fitted model.

[Package NMF version 0.21.0 Index]