bterms<- {NMF}	R Documentation

Fixed Terms in NMF Models

Description

These functions are for internal use and should not be called by the end-user.

cterms<- sets fixed coefficient terms or indexes and should only be called on a newly created NMF object, i.e. in the constructor/factory generic nmfModel.

Usage

  bterms(object)<-value

  cterms(object)<-value

Arguments

object
NMF object to be updated.

value
specification of the replacement value for fixed-terms.

Details

They use model.matrix(~ -1 + ., data=value) to generate suitable term matrices.

Methods

bterms<-

signature(object = "NMFstd"): Default method tries to coerce value into a data.frame with as.data.frame.

cterms<-

signature(object = "NMFstd"): Default method tries to coerce value into a data.frame with as.data.frame.

[Package NMF version 0.21.0 Index]