txtProgressBar {NMF}	R Documentation

Simple Progress Bar

Description

Creates a simple progress bar with title. This function is identical to utils::txtProgressBar but allow adding a title to the progress bar, and can be shared by multiple processes, e.g., in multicore or multi-hosts computations.

Usage

  txtProgressBar(min = 0, max = 1, initial = 0, char = "=",
    width = NA, title = if (style == 3) " ", label,
    style = 1, file = "", shared = NULL)

Arguments

shared
specification of a shared directory to use when the progress bar is to be used by multiple processes.

min
(finite) numeric values for the extremes of the progress bar. Must have min < max.

max
(finite) numeric values for the extremes of the progress bar. Must have min < max.

initial
initial or new value for the progress bar. See ‘Details’ for what happens with invalid values.

char
the character (or character string) to form the progress bar.

width
the width of the progress bar, as a multiple of the width of char. If NA, the default, the number of characters is that which fits into getOption("width").

title
ignored, for compatibility with other progress bars.

label
ignored, for compatibility with other progress bars.

style
the ‘style’ of the bar – see ‘Details’.

file
an open connection object or "" which indicates the console: stderr() might be useful here.

Author(s)

R Core Team

[Package NMF version 0.21.0 Index]