utils-NMF {NMF}	R Documentation

Utility Function in the NMF Package

Description

Utility Function in the NMF Package

str_args formats the arguments of a function using args, but returns the output as a string.

Usage

  str_args(x, exdent = 10L)

Arguments

x
a function

exdent
indentation for extra lines if the output takes more than one line.

Examples

args(library)
str_args(library)

[Package NMF version 0.21.0 Index]